requirements.txt
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
choreographer==1.0.10
click==8.2.1
constantly==23.10.4
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
daphne==4.2.1
decorator==5.2.1
Django==5.2.6
django-channels==0.7.0
django-cors-headers==4.8.0
djangorestframework==3.16.1
executing==2.2.0
filelock==3.19.1
filetype==1.2.0
flasgger==0.9.7.1
Flask==3.1.2
flask-sock==0.7.0
fonttools==4.59.1
fsspec==2025.7.0
greenlet==3.2.4
h11==0.16.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kaleido==1.0.0
kiwisolver==1.4.9
lap==0.5.12
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.1.4
mpmath==1.3.0
msgpack==1.1.1
narwhals==2.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
plotly==6.3.0
prompt_toolkit==3.0.52
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
roboflow==1.2.7
rpds-py==0.27.1
scipy==1.16.1
service-identity==24.2.0
setuptools==80.9.0
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.183
ultralytics-thop==2.0.16
urllib3==2.5.0
vanna==0.7.9
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
zope.interface==8.0
